New Design for our page
Flow of the page:
UI: The movie titles will be in the left column. They will be buttons for the users to click. When a user selects a movie title, the movies ReView information is displayed in the large container in the middle. The corresponding movie information will be displayed in the box as shown. 

Back End Data: The titles will have three variables, name, review and rating.